+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dropped `net6.0` support
 - Split directory for code coverage and code coverage history between unit and integration tests for both.
 - `IPushNuGetPackages` component no longer add `--skip-duplicate` option by default.
+- Replaced `Configure<DotNetRunSettings> StrykerArgumentSettings` is now `Configure<StrykerOptions>`
+
 
 ### 🛠️ Fixes
 - `IBuildDockerImage` component should run AFTER `IUnitTest`, `IIntegrationTest` and `IMutationTest` components.